feat: support extract LoginState username, log it
First steps towards being more content-aware. We are now able to get the
unauthenticated LoginState username and print it out in the logs to
determime who started the server (or initiated a login event). Future
work is being done to make this be post-login stage so we are able to
guarantee (online mode) that a user is who they said they are.

This also starts to lay the groundwork for an event bus to handle
notifications, but that is much farther away.

Also added a `make reload` target to make it easier to develop.