feat: initial eventbus implementation
Adds an `eventbus` package meant to make it easy to react to events
synchronously. This will be used for chat integration as well as other
places later on. Added tests for the eventbus package and started work
on hooking it up into the proxy layer.

This will be considered complete when I move the whitelist system to use
it. We will also potentially have a plugin system built into this to
support these events over the wire instead of just in-memory.