feat: use a single listener
Instead of using a listener per-server, we now use a single global
listener. This listener defaults to listening on `0.0.0.0:25565`, like
the previous per-server listener did before.

As part of this, the configuration objects have been changed. `name` has
been changed to `hostname`. This hostname should match exactly what the
Minecraft client is using to connect. It is self-reported by the client
during the handshake process[^1]. The example configuration has been
updated to match the new format, as well as the docs. I also added a
WSL-specific `listenAddress` hint to help with port-forwarding issues on
WSL machines.

This is a BREAKING CHANGE because of the configuration changes required,
so I will be releasing a v3 post-merge.

[^1]: https://wiki.vg/Protocol#Handshaking